refactor: organize story sidebar components into dedicated story folder
- Create src/components/sidebar/story/ folder for story-specific components
- Move StorySidebar, StoriesSection, InfoSection, VocabularySection to story folder
- Update all import paths to use direct file imports (no barrel exports)
- Remove barrel export index.ts files per project guidelines
- Update imports in pages and test files to use new folder structure
- Maintain all existing functionality with improved organization
