--- src/components/sidebar/base/index.ts
@@ -1,2 +1,0 @@
-export { default as BaseSidebar } from './BaseSidebar';
-export { default as BaseSidebarHeader } from './BaseSidebarHeader';
