+++ src/components/sidebar/story/InfoSection.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { Card, CardContent, CardHeader, CardTitle } from '../ui/Card';
-import { Badge } from '../ui/Badge';
-import Label from '../ui/Label';
+import { Card, CardContent, CardHeader, CardTitle } from '../../ui/Card';
+import { Badge } from '../../ui/Badge';
+import Label from '../../ui/Label';
 import type { TFunction } from 'i18next';
-import type { DifficultyLevel } from '../../types/llm/prompts';
-import type { TranslationResponse } from '../../lib/translationService';
-import { cn } from '../../lib/utils';
-import type { LanguageCode } from '../../types/llm/prompts';
+import type { DifficultyLevel } from '../../../types/llm/prompts';
+import type { TranslationResponse } from '../../../lib/translationService';
+import { cn } from '../../../lib/utils';
+import type { LanguageCode } from '../../../types/llm/prompts';
 import { CheckCircle, AlertTriangle } from 'lucide-react';
 
 interface InfoSectionProps {
