+++ src/components/sidebar/story/StorySidebar.tsx
@@ -15,7 +15,8 @@ import { TokenConverter } from '../../lib/llm/tokens/tokenConverter';
 import { Button } from '../ui/Button';
 import { BookOpen, BookMarked, Settings } from 'lucide-react';
 
-import { BaseSidebar, BaseSidebarHeader } from './base';
+import BaseSidebar from '../base/BaseSidebar';
+import BaseSidebarHeader from '../base/BaseSidebarHeader';
 import StoriesSection from './StoriesSection';
 import VocabularySection from './VocabularySection';
 import InfoSection from './InfoSection';
